+++ test_iko.py
@@ -1,19 +1,46 @@
+import base64
+
 import pytest
 
 import iko
 
 
-class Human(iko.Schema):
+CONTEXT = {'private_offset': 1}
+
+
+class PrivateField(iko.Field):
+    async def post_load(self, value, context):
+        value = bytes(
+            (char + context['private_offset']) % 256
+            for char in value.encode('utf-8')
+        )
+        return base64.encodebytes(value).decode('utf-8').strip()
+
+    async def post_dump(self, value, context):
+        value = bytes(
+            (bt - context['private_offset']) % 256
+            for bt in base64.decodebytes(value.encode('utf-8'))
+        )
+        return value.decode('utf-8').strip()
+
+
+class Named(iko.Schema):
     name = iko.Field()
+
+
+class HumanSchema(Named):
     gender = iko.Field()
+    secret = PrivateField()
 
 
-class CatSchema(iko.Schema):
-    _id = iko.Field(outer_name='id')
-    nickname = iko.Field()
+class Pet(iko.Schema):
+    owner = iko.Nested(HumanSchema)
+    friends = iko.List(HumanSchema)
     age = iko.Field(default='infinity')
-    owner = iko.Nested(Human)
-    friends = iko.List(Human)
+
+
+class CatSchema(Named, Pet):
+    _id = iko.Field(outer_name='id')
     skip = iko.Field()
 
 
@@ -23,10 +50,11 @@ class CatSchema(iko.Schema):
         (
             {
                 '_id': '42',
-                'nickname': 'Jake',
+                'name': 'Jake',
                 'owner': {
                     'name': 'Ilya',
-                    'gender': 'male'
+                    'gender': 'male',
+                    'secret': 'bnBwbw=='
                 },
                 'friends': [
                     {
@@ -41,11 +69,12 @@ class CatSchema(iko.Schema):
             },
             {
                 'id': '42',
-                'nickname': 'Jake',
+                'name': 'Jake',
                 'age': 'infinity',
                 'owner': {
                     'name': 'Ilya',
-                    'gender': 'male'
+                    'gender': 'male',
+                    'secret': 'moon'
                 },
                 'friends': [
                     {
@@ -58,11 +87,18 @@ class CatSchema(iko.Schema):
                     }
                 ]
             }
+        ),
+        (
+            {},
+            {
+                'age': 'infinity',
+            }
         )
     ]
 )
 async def test_dump(data, expected):
-    assert await CatSchema.dump(data) == expected
+    assert await CatSchema.dump(data, context=CONTEXT) == expected
+    assert await CatSchema.dump_many([data], context=CONTEXT) == [expected]
 
 
 @pytest.mark.parametrize(
@@ -71,10 +107,11 @@ async def test_dump(data, expected):
         (
             {
                 'id': '42',
-                'nickname': 'Jake',
+                'name': 'Jake',
                 'owner': {
                     'name': 'Ilya',
-                    'gender': 'male'
+                    'gender': 'male',
+                    'secret': 'moon'
                 },
                 'friends': [
                     {
@@ -89,11 +126,12 @@ async def test_dump(data, expected):
             },
             {
                 '_id': '42',
-                'nickname': 'Jake',
+                'name': 'Jake',
                 'age': 'infinity',
                 'owner': {
                     'name': 'Ilya',
-                    'gender': 'male'
+                    'gender': 'male',
+                    'secret': 'bnBwbw=='
                 },
                 'friends': [
                     {
@@ -106,8 +144,33 @@ async def test_dump(data, expected):
                     }
                 ]
             }
+        ),
+        (
+            {},
+            {
+                'age': 'infinity',
+            }
         )
     ]
 )
 async def test_load(data, expected):
+    assert await CatSchema.load(data, context=CONTEXT) == expected
+    assert await CatSchema.load_many([data], context=CONTEXT) == [expected]
+
+
+@pytest.mark.parametrize(
+    'data, expected',
+    [
+        (
+            {},
+            {
+                'age': 'infinity',
+            }
+        )
+    ]
+)
+async def test_empty_context(data, expected):
+    assert await CatSchema.dump(data) == expected
+    assert await CatSchema.dump_many([data]) == [expected]
     assert await CatSchema.load(data) == expected
+    assert await CatSchema.load_many([data]) == [expected]