+++ CHANGELOG.rst
@@ -3,12 +3,19 @@ Changelog
 
 A list of changes between each release.
 
-0.1.0 (2019-10-18)
+0.3.0 (2019-11-14)
 ^^^^^^^^^^^^^^^^^^
 
-- First release
+- feat: add many_dump/many_load methods
+- feat: default dict in context
+- bug: add post_load method to Nested schema
 
 0.2.0 (2019-10-28)
 ^^^^^^^^^^^^^^^^^^
 
 - feat: add post methods and use gather
+
+0.1.0 (2019-10-18)
+^^^^^^^^^^^^^^^^^^
+
+- First release